line up names with server types
